refactor: 完成多项代码优化与功能完善
- 重构日志统计与token处理逻辑，使用int64避免溢出问题
- 更新默认日志保留天数从7天改为30天
- 修复前端组件属性不兼容问题（width→size、direction→orientation）
- 新增ChartCard、ChartEmpty、useShallowMemo等前端组件与工具
- 新增应用生命周期与缓存测试，完善数据持久化逻辑
- 优化Windows托盘代码，移除无用的PeekMessageW调用
- 更新依赖包版本，新增sqlite与相关工具库
- 为Anthropic流式响应添加token统计支持

+++ internal/server/rewrite_test.go
@@ -87,8 +87,8 @@ func TestIsStreaming(t *testing.T) {
 func TestParseUsage(t *testing.T) {
 	cases := []struct {
 		body    string
-		wantIn  int
-		wantOut int
+		wantIn  int64
+		wantOut int64
 	}{
 		{`{"usage":{"prompt_tokens":11,"completion_tokens":22}}`, 11, 22},
 		{`{}`, 0, 0},
